Specialized autoscale_parabols function for precise scale coefficient. + some little changes
If k=2n+1, n is integer, then medium parabol's point is not match to
input meaning.

Only 6 forming points are considered.

+++ ruledparabols.cpp
@@ -31,7 +31,7 @@ double *RuledParabols::getPoint(double u, double w)
     return point;
 }
 
-Figure* RuledParabols::getFigure(double k)
+Figure* RuledParabols::getFigure(int k)
 {
     double step = 1.0/k;
     double i0 = 0;
